+++ hatch/cli/version/__init__.py
@@ -6,6 +6,13 @@ import click
 @click.pass_obj
 def version(app, desired_version):
     """View or set a project's version."""
+    if app.project.metadata.core.version is not None:
+        if desired_version:
+            app.abort('Cannot set version when it is statically defined by the `project.version` field')
+        else:
+            app.display_info(app.project.metadata.core.version)
+            return
+
     source = app.project.config.version.source
 
     version_data = source.get_version_data()